2.intermediate/requirements.txt
uvicorn[standard]
fastapi
bcrypt
python-dotenv
python-jose[cryptography]
python-multipart
